Move previewExam and createOfflineExam to separate files
